+++ docs/resources/changelog.md
@@ -18,6 +18,7 @@ Technical reasons aside, canceling these events seldom led to a good user experi
 - Added Iconoir example to the icon docs
 - Changed the `cancelable` default to `false` for the internal `@event` decorator
 - Fixed a bug where toggling `open` stopped working in `sl-alert`, `sl-dialog`, `sl-drawer`, `sl-dropdown`, and `sl-tooltip`
+- Fixed a bug in `sl-range` where setting a value below `min` would cause the value to equal `min` [#448](https://github.com/shoelace-style/shoelace/issues/448)
 - Fixed a number of imports that should have been type imports
 - Updated Lit to 2.0.0-rc.2
 - Updated esbuild to 0.12.4